requirements.txt
fastapi
uvicorn
python-binance
python-dotenv
streamlit
mysql-connector-python
pyjwt
websockets